moved components and reducers into common (ssr/front) directory.
